feat: split session.ts into focused modules below 400-line limit (US-000)
