+++ .nax/status.json
@@ -30,7 +30,7 @@
     "phase": "routing"
   },
   "iterations": 1,
-  "updatedAt": "2026-04-04T05:13:54.700Z",
-  "durationMs": 1260066,
-  "lastHeartbeat": "2026-04-04T05:13:54.700Z"
+  "updatedAt": "2026-04-04T05:21:54.666Z",
+  "durationMs": 1740032,
+  "lastHeartbeat": "2026-04-04T05:21:54.666Z"
 }
