+++ src/debate/index.ts
@@ -2,8 +2,9 @@
  * Debate module barrel export
  */
 
-export { DebateSession, _debateSessionDeps } from "./session";
-export type { DebateSessionOptions } from "./session";
+export { DebateSession } from "./session";
+export { _debateSessionDeps, resolveDebaterModel } from "./session-helpers";
+export type { DebateSessionOptions } from "./session-helpers";
 export { majorityResolver, synthesisResolver, judgeResolver } from "./resolvers";
 export { buildCritiquePrompt, buildSynthesisPrompt, buildJudgePrompt } from "./prompts";
 export type {
